Ver2_USING_THIS/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
numpy==1.23.5
paho-mqtt==1.6.1
pandas==1.5.2
psycopg2==2.9.3
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.4
djangorestframework-simplejwt

